+++ array/docs/types/test.ts
@@ -1,4 +1,4 @@
-/**
+/*
 * @license Apache-2.0
 *
 * Copyright (c) 2025 The Stdlib Authors.
@@ -16,32 +16,14 @@
 * limitations under the License.
 */
 
-'use strict';
+/* eslint-disable @typescript-eslint/no-unused-expressions */
 
-var abs = require( './../../../base/special/abs' );
-var ns = require( './../lib' );
+import ns = require( './index' );
 
-// Define a list of supported input dtypes:
-var idtypes = [
-	'float64',
-	'float32',
-	'generic'
-];
 
-// Define a list of supported output dtypes:
-var odtypes = [
-	'float64',
-	'float32',
-	'generic'
-];
+// TESTS //
 
-// Create a function for applying a unary function to each element of an array:
-var f = new ns.unary( abs, idtypes, odtypes, 'same' );
-
-// Create an input array:
-var x = [ -1.0, 2.0, -3.0, 4.0 ];
-
-// Perform element-wise computation:
-var out = f.apply( x );
-console.log( out );
-// => [ 1.0, 2.0, 3.0, 4.0 ]
+// The exported value is the expected interface...
+{
+	ns; // $ExpectType Namespace
+}